commands: make pages consistent in design (#9)
Make the command pages consistent in their design of information.

The pages all start with their usage and supported key types, followed
by a summary of the command. A section with a list of possible errors
follows, and then a section with examples. One or two example
sub-sections are included using the CLI and Rust client.

Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>